requirements.txt
Django>=5.0,<6.0
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
django-cors-headers
celery